Backend/domi_foods/requirements.txt
django-rest-framework
django-rest-authtoken
django-rest-knox
Pillow
knox
django-cors-headers
